+++ lib/castle/configuration.rb
@@ -22,9 +22,9 @@ module Castle
       \Aunix:
     /ix].freeze
 
-    # @note this value is not assigned as we don't recommend using a whitelist. If you need to use
+    # @note this value is not assigned as we don't recommend using a allowlist. If you need to use
     #   one, this constant is provided as a good default.
-    DEFAULT_WHITELIST = %w[
+    DEFAULT_ALLOWLIST = %w[
       Accept
       Accept-Charset
       Accept-Datetime
@@ -44,8 +44,7 @@ module Castle
     ].freeze
 
     attr_accessor  :request_timeout, :url, :trust_proxy_chain
-    attr_reader :api_secret, :whitelisted, :blacklisted, :failover_strategy, :ip_headers,
-                :trusted_proxies, :trusted_proxy_depth
+    attr_reader :api_secret, :allowlisted, :denylisted, :failover_strategy, :ip_headers,
 
     def initialize
       @formatter = Castle::HeadersFormatter
@@ -56,8 +55,8 @@ module Castle
     def reset
       self.failover_strategy = FAILOVER_STRATEGY
       self.url = URL
-      self.whitelisted = [].freeze
-      self.blacklisted = [].freeze
+      self.allowlisted = [].freeze
+      self.denylisted = [].freeze
       self.api_secret = ENV.fetch('CASTLE_API_SECRET', '')
       self.ip_headers = [].freeze
       self.trusted_proxies = [].freeze
@@ -73,12 +72,12 @@ module Castle
       @api_secret = value.to_s
     end
 
-    def whitelisted=(value)
-      @whitelisted = (value ? value.map { |header| @formatter.call(header) } : []).freeze
+    def allowlisted=(value)
+      @allowlisted = (value ? value.map { |header| @formatter.call(header) } : []).freeze
     end
 
-    def blacklisted=(value)
-      @blacklisted = (value ? value.map { |header| @formatter.call(header) } : []).freeze
+    def denylisted=(value)
+      @denylisted = (value ? value.map { |header| @formatter.call(header) } : []).freeze
     end
 
     # sets ip headers